+++ cacus/duploader.py
@@ -4,33 +4,36 @@
 import os
 import time
 import atexit
-import logging
 import pyinotify
 import threading
+import Queue
 from binascii import hexlify
 from debian import deb822
 
-import repo_manage
 import common
+import repo_manage
 
-log = logging.getLogger('cacus.duploader')
 
+class ComplexDistroWatcher(pyinotify.ProcessEvent):
+    """ inotify watcher with full set of features.
+    Supports including .changes file signature check, uploading of package sources etc.
+    Also can upload single .deb binary packages.
+    """
 
-class EventHandler(pyinotify.ProcessEvent):
-
-    def __init__(self, settings, component):
+    def __init__(self, repo_manager, settings, component):
+        self.repo_manager = repo_manager
         self.distro = settings['distro']
         self.component = component
         self.gpg_check = settings.get('gpg_check', True)
         self.strict = settings.get('strict', True)
         self.incoming_wait_timeout = settings.get('incoming_wait_timeout', 5)
-        self.log = logging.getLogger('cacus.duploader.{0}'.format(self.distro))
+        self.log = repo_manager.log
         self.uploaded_files = {}
         self.uploaded_files_lock = threading.Lock()
         self.uploaded_event = threading.Event()
 
     def _gpgCheck(self, data):
-        result = common.gpg.verify(data)
+        result = self.repo_manager.gpg.verify(data)
         if not result.valid:
             if result.status:
                 raise Exception("signed with {}, status '{}'".format(result.key_id, result.status))
@@ -38,15 +41,16 @@ class EventHandler(pyinotify.ProcessEvent):
                 raise Exception("Bad PGP data")
         return result.username
 
-    def _process_single_deb(self, distro, component, file):
+    def _process_single_deb(self, file):
         if os.path.isfile(file) and file in self.uploaded_files:
             # if file is still exists and wasn't picked by some _processChangesFile(),
             # assume that it was meant to be uploaded as signle package
             with self.uploaded_files_lock:
                 self.uploaded_files.pop(file)
-            self.log.debug("Uploading %s to %s/%s", file, distro, component)
+            self.log.debug("Uploading %s to %s/%s", file, self.distro, self.component)
             try:
-                common.with_retries(repo_manage.upload_package, distro, component, [file], changes=None, forceUpdateMeta=True)
+                common.with_retries(self.repo_manager.config['retry_count'], self.repo_manager.config['retry_delays'],
+                                    self.repo_manager.upload_package, self.distro, self.component, [file], changes=None)
             except Exception as e:
                 self.log.error("Error while uploading DEB %s: %s", file, e)
             os.unlink(file)
@@ -118,8 +122,9 @@ class EventHandler(pyinotify.ProcessEvent):
             self.log.info("%s-%s: sign: OK, checksums: OK, uploading to distro '%s', component '%s'",
                           changes['source'], changes['version'], self.distro, self.component)
             try:
-                common.with_retries(repo_manage.upload_package, self.distro, self.component, incoming_files, changes=changes, forceUpdateMeta=True)
-                self.log.warn("OK")
+                common.with_retries(self.repo_manager.config['retry_count'], self.repo_manager.config['retry_delays'],
+                                    self.repo_manager.upload_package, self.distro, self.component,
+                                    incoming_files, changes=changes, forceUpdateMeta=True)
             except Exception as e:
                 self.log.error("Error while uploading file: %s", e)
 
@@ -134,7 +139,7 @@ class EventHandler(pyinotify.ProcessEvent):
             thread.start()
         else:
             # store uploaded file and send event to all waiting threads
-            self.uploaded_files[event.pathname] = common.get_hashes(filename=event.pathname)
+            self.uploaded_files[event.pathname] = self.repo_manager.get_hashes(filename=event.pathname)
             self.uploaded_event.set()
             self.uploaded_event.clear()
 
@@ -142,57 +147,118 @@ class EventHandler(pyinotify.ProcessEvent):
         # so schedule uploader worker after 2*incoming timeout (i.e. deb was not picked by _processChangesFile)
         if not self.strict and (event.pathname.endswith('.deb') or event.pathname.endswith('.udeb')):
             self.log.info("Will upload it within %s seconds", 2*self.incoming_wait_timeout)
-            uploader = threading.Timer(2*self.incoming_wait_timeout, self._process_single_deb, args=(self.distro, self.component, event.pathname))
+            uploader = threading.Timer(2*self.incoming_wait_timeout, self._process_single_deb, args=(event.pathname,))
             uploader.daemon = True
             uploader.start()
 
 
-def _cleanup(watchers):
-    for w, notifier in watchers.items():
-        notifier.stop()
-        del watchers[w]
+class SimpleDistroWatcher(pyinotify.ProcessEvent):
+    """ inotify watcher for simple repos
+    Supports simple, binary-only distibutions, which though is enough for most of users.
+    Faster than ComplexDistroWatcher because packages are uploaded immidiately, without waiting for .changes file
+    """
 
+    def __init__(self, repo_manager, settings, component):
+        self.repo_manager = repo_manager
+        self.distro = settings['distro']
+        self.component = component
+        self.log = repo_manager.log
+        self.queue = Queue.Queue()
+        self.worker = threading.Thread(target=self._worker)
+        self.worker.daemon = True
+        self.worker.start()
 
-def start_duploader():
-    watchers = {}
-    atexit.register(_cleanup, watchers)
-    incoming_root = common.config['duploader_daemon']['incoming_root']
-    if not os.path.isdir(incoming_root):
-        os.mkdir(incoming_root)
-
-    try:
+    def _worker(self):
+        count = 0
         while True:
-            # check out for any new distros in DB (except read-only snapshots) and create watchers for each component of distro, if any
-            distros = list(common.db_cacus.distros.find({'snapshot': {'$exists': False}, 'imported': {'$exists': False}}))
-            for distro_settings in distros:
-                distro = distro_settings['distro']
-                if distro not in watchers:
-                    watchers[distro] = {}
+            file = self.queue.get()
+            self.log.debug("Uploading %s to %s/%s", file, self.distro, self.component)
+            try:
+                common.with_retries(self.repo_manager.config['retry_count'], self.repo_manager.config['retry_delays'],
+                                    self.repo_manager.upload_package, self.distro, self.component, [file], changes=None, skipUpdateMeta=True)
+                count += 1
+            except Exception as e:
+                self.log.error("Error while uploading DEB %s: %s", file, e)
+            os.unlink(file)
+            if self.queue.empty() or count >= 10:
+                count = 0
+                self.repo_manager.update_distro_metadata(self.distro, comps=[self.component])
 
-                components = [x['component'] for x in common.db_cacus.components.find({'distro': distro})]
-                for comp in components:
-                    if comp not in watchers[distro]:
-                        incoming_dir = os.path.join(incoming_root, distro, comp)
-                        try:
-                            if not os.path.isdir(incoming_dir):
-                                log.debug("Creating incoming dir '%s'", incoming_dir)
-                                os.makedirs(incoming_dir, mode=0777)
-                        except Exception as e:
-                            log.error("Cannot create dir for incoming files '%s': %s", incoming_dir, e)
-                            continue
-                        handler = EventHandler(distro_settings, comp)
-                        wm = pyinotify.WatchManager()
-                        notifier = pyinotify.ThreadedNotifier(wm, handler)
-                        wm.add_watch(incoming_dir, pyinotify.ALL_EVENTS)
-                        log.info("Starting notifier for '%s/%s' at %s, strict: %s", distro, comp, incoming_dir, distro_settings['strict'])
-                        notifier.start()
-                        watchers[distro][comp] = notifier
+    def process_IN_CLOSE_WRITE(self, event):
+        self.log.info("Got file %s", event.pathname)
+        if event.pathname.endswith('.deb') or event.pathname.endswith('.udeb'):
+            self.queue.put(event.pathname)
+        else:
+            os.unlink(event.pathname)
 
-                abandoned = set(watchers[distro].keys()) - set(components)
-                for comp in abandoned:
-                    log.info("Removing notifier for '%s/%s'", distro, comp)
-                    watchers[distro][comp].stop()
-                    del(watchers[distro][comp])
-            time.sleep(5)
-    except KeyboardInterrupt:
-        _cleanup(watchers)
+
+class Duploader(repo_manage.RepoManager):
+
+    def _cleanup(self, watchers):
+        for distro, distro_watchers in watchers.items():
+            for comp, notifier in distro_watchers.items():
+                self.log.info("Removing notifier for '%s/%s'", distro, comp)
+                notifier.stop()
+                del distro_watchers[comp]
+            del watchers[distro]
+
+    def run(self):
+        self.watchers = {}
+        atexit.register(self._cleanup, self.watchers)
+        incoming_root = self.config['duploader_daemon']['incoming_root']
+        if not os.path.isdir(incoming_root):
+            os.mkdir(incoming_root)
+
+        try:
+            while True:
+                # check out for any new distros in DB (except read-only snapshots) and create watchers for each component of distro, if any
+                distros = list(self.db.cacus.distros.find({'snapshot': {'$exists': False}, 'imported': {'$exists': False}}))
+                for distro_settings in distros:
+                    distro = distro_settings['distro']
+                    if distro not in self.watchers:
+                        self.watchers[distro] = {}
+
+                    components = [x['component'] for x in self.db.cacus.components.find({'distro': distro})]
+                    for comp in components:
+                        if comp not in self.watchers[distro]:
+                            incoming_dir = os.path.join(incoming_root, distro, comp)
+                            try:
+                                if not os.path.isdir(incoming_dir):
+                                    self.log.debug("Creating incoming dir '%s'", incoming_dir)
+                                    os.makedirs(incoming_dir, mode=0777)
+                            except Exception as e:
+                                self.log.error("Cannot create dir for incoming files '%s': %s", incoming_dir, e)
+                                continue
+
+                            if distro_settings.get('simple', True):
+                                handler = SimpleDistroWatcher(self, distro_settings, comp)
+                            else:
+                                handler = ComplexDistroWatcher(self, distro_settings, comp)
+                            wm = pyinotify.WatchManager()
+                            notifier = pyinotify.ThreadedNotifier(wm, handler)
+                            wm.add_watch(incoming_dir, pyinotify.ALL_EVENTS)
+                            self.log.info("Starting %s for '%s/%s' at %s, strict: %s",
+                                          type(handler).__name__, distro, comp, incoming_dir, distro_settings['strict'])
+                            notifier.start()
+                            self.watchers[distro][comp] = notifier
+
+                    abandoned = set(self.watchers[distro].keys()) - set(components)
+                    for comp in abandoned:
+                        self.log.info("Removing notifier for '%s/%s'", distro, comp)
+                        self.watchers[distro][comp].stop()
+                        del(self.watchers[distro][comp])
+
+                for distro in set(self.watchers.keys()) - set(x['distro'] for x in distros):
+                    for comp in self.watchers[distro]:
+                        self.log.info("Removing notifier for '%s/%s'", distro, comp)
+                        self.watchers[distro][comp].stop()
+                    del(self.watchers[distro])
+
+                time.sleep(5)
+        except KeyboardInterrupt:
+            self._cleanup(self.watchers)
+
+
+def start_daemon(config):
+    duploader = Duploader(config)
+    duploader.run()