+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## [Version 0.6](https://github.com/beebeeep/cacus/tree/v0.6) (2017-03-13)
+  * Refactored code
+  * Added "simple" duploader - for binary-only repos
+  * Added distro removal
+
 ## [Version 0.5](https://github.com/beebeeep/cacus/tree/v0.5) (2017-03-08)
   * Added incoming dirs for each component of each Distro
   * Distro components list is now fixed during repo creation and can be only
     changed by updating distro settings via API
-